draw: drop 8-bit single-element ops, write 16-bit cov tri-state
Reverts 2f0ac51a + 4b3cd54e (op_load_8 / op_store_8 / op_gather_8 plus
the umbra_{load,store,gather}_8 public API) and switches the SDF bounds
program to write a 16-bit cov tri-state via umbra_store_16.  Coverage
glyph samplers (coverage.c) and their tests follow suit, storing
uint16_t bitmaps and using umbra_load_16 / umbra_gather_16 with
umbra_i32_from_u16.

The 8-bit ops were never bit-exact across backends.  spirv.c lowered
op_store_8 as a non-atomic read-modify-write of the containing u32 word,
which races between threads writing different bytes within one word -- a
limitation that was tolerable while the only writer (the SDF cov shader)
ran exclusively on JIT/interp.  Running the bounds program on a GPU
backend exposes the race; reverting the ops sidesteps it without giving
up store_8x4 / load_8x4 (32-bit-aligned 4-byte writes are race-free).

Keeps load_8x4 / store_8x4 (32-bit RGBA pack/unpack), which never had
this problem.  Updates dumps/ accordingly.

+++ tests/draw_test.c
@@ -769,7 +769,7 @@ TEST(test_coverage_rect_white_dst) {
 TEST(test_coverage_bitmap) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 1, 1, 1};
-    uint8_t  cov_data[8] = {0, 128, 255, 0, 0, 0, 0, 0};
+    uint16_t cov_data[8] = {0, 128, 255, 0, 0, 0, 0, 0};
     struct umbra_buf buf = {.ptr=cov_data, .count=8};
     struct umbra_builder *builder = umbra_builder();
     umbra_ptr const dst_ptr = umbra_bind_buf(builder, &dst_slot);
@@ -798,7 +798,7 @@ TEST(test_coverage_bitmap) {
 TEST(test_coverage_bitmap_565) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 0, 0, 1};
-    uint8_t  cov_data[16];
+    uint16_t cov_data[16];
     for (int i = 0; i < 16; i++) { cov_data[i] = 255; }
     struct umbra_buf buf = {.ptr=cov_data, .count=16};
     struct umbra_builder *builder = umbra_builder();
@@ -827,7 +827,7 @@ TEST(test_coverage_bitmap_565) {
 TEST(test_coverage_sdf) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 1, 1, 1};
-    uint8_t  cov_data[8] = {0, 100, 128, 200, 255, 0, 0, 0};
+    uint16_t cov_data[8] = {0, 100, 128, 200, 255, 0, 0, 0};
     struct umbra_buf buf = {.ptr=cov_data, .count=8};
     struct umbra_builder *builder = umbra_builder();
     umbra_ptr const dst_ptr = umbra_bind_buf(builder, &dst_slot);
@@ -854,7 +854,7 @@ TEST(test_coverage_sdf) {
 TEST(test_coverage_bitmap_matrix) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 1, 1, 1};
-    uint8_t  bmp[8] = {0, 0, 255, 0, 0, 0, 0, 0};
+    uint16_t bmp[8] = {0, 0, 255, 0, 0, 0, 0, 0};
     struct coverage_bitmap2d sampler = {
         .buf = {.ptr=bmp, .count=8}, .w=8, .h=1,
     };
@@ -886,7 +886,7 @@ TEST(test_coverage_bitmap_matrix) {
 TEST(test_coverage_bitmap_matrix_565) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 0, 0, 1};
-    uint8_t  bmp[16];
+    uint16_t bmp[16];
     for (int i = 0; i < 16; i++) { bmp[i] = 255; }
     struct coverage_bitmap2d sampler = {
         .buf = {.ptr=bmp, .count=16}, .w=16, .h=1,
@@ -919,7 +919,7 @@ TEST(test_coverage_bitmap_matrix_565) {
 TEST(test_coverage_bitmap_matrix_oob) {
     struct umbra_buf dst_slot = {0};
     umbra_color color = {1, 1, 1, 1};
-    uint8_t  bmp[4] = {255, 0, 0, 0};
+    uint16_t bmp[4] = {255, 0, 0, 0};
     struct coverage_bitmap2d sampler = {
         .buf = {.ptr=bmp, .count=4}, .w=2, .h=2,
     };
